Distributing AVPickupDropoffTest over multiple files
