+++ src/test/java/ch/ethz/matsim/av/dynamics/TestRequest.java
@@ -1,0 +1,19 @@
+package ch.ethz.matsim.av.dynamics;
+
+/**
+ * Creates a new TestRequest
+ */
+public class TestRequest {
+    public final double departureTime;
+    public final double delayTime;
+
+    /**
+     * Constructor
+     * @param departureTime
+     * @param delayTime
+     */
+    public TestRequest(double departureTime, double delayTime) {
+        this.departureTime = departureTime;
+        this.delayTime = delayTime;
+    }
+}
